mnemonic_shift.cpp -> mnemonic_calc.cpp
