+++ src/mnemonic.cpp
@@ -5,6 +5,7 @@ void mnemonic_bit_op(LineData* line, Mnemonic mne);                  // mnemonic
 void mnemonic_calc8(LineData* line, uint8_t code);                   // mnemonic_calc.cpp
 void mnemonic_calc16(LineData* line, uint8_t code);                  // mnemonic_calc.cpp
 void mnemonic_calcOH(LineData* line, uint8_t code8, uint8_t code16); // mnemonic_calc.cpp
+void mnemonic_shift(LineData* line, uint8_t code);                   // mnemonic_calc.cpp
 void mnemonic_CALL(LineData* line);                                  // mnemonic_call.cpp
 void mnemonic_RET(LineData* line);                                   // mnemonic_call.cpp
 void mnemonic_RST(LineData* line);                                   // mnemonic_call.cpp
@@ -17,7 +18,6 @@ void mnemonic_JP(LineData* line);                                    // mnemonic
 void mnemonic_JR(LineData* line);                                    // mnemonic_jump.cpp
 void mnemonic_DJNZ(LineData* line);                                  // mnemonic_jump.cpp
 void mnemonic_LD(LineData* line);                                    // mnemonic_load.cpp
-void mnemonic_shift(LineData* line, uint8_t code);                   // mnemonic_shift.cpp
 void mnemonic_PUSH(LineData* line);                                  // mnemonic_stack.cpp
 void mnemonic_POP(LineData* line);                                   // mnemonic_stack.cpp
 void mnemonic_DB(LineData* line);                                    // mnemonic_data.cpp
